requirements.txt
flask
gunicorn
python-pptx
pillow